(Backend) Move libappearancecell.h out of public
