+++ Makefile
@@ -13,8 +13,8 @@ libappearancecell_PRIVATE_FRAMEWORKS = Preferences
 after-all::
 	cp .theos/obj/debug/libappearancecell.dylib $(THEOS)/lib
 	mkdir -p $(THEOS_STAGING_DIR)/usr/include/libappearancecell
-	cp ./public/*.h $(THEOS_STAGING_DIR)/usr/include/libappearancecell
-	cp -r ./public/*.h $(THEOS)/include/libappearancecell/libappearancecell
+	cp libappearancecell.h $(THEOS_STAGING_DIR)/usr/include/libappearancecell
+	cp libappearancecell.h $(THEOS)/include/libappearancecell/libappearancecell
 
 # SUBPROJECTS += PreferencesExample
