+++ libappearancecell.h
